backend/requirement.txt
content_filtering_projectFlask==2.2.2
joblib==1.2.0
pandas==1.5.3
scikit-learn==1.2.2
